refactor layout to include header and footer
