+++ src/components/Footer.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+export default function Footer() {
+  return (
+    <footer className="w-full py-4 text-center text-sm text-sage-600">
+      &copy; {new Date().getFullYear()} Aglamaz
+    </footer>
+  );
+}
