+++ src/components/Header.tsx
@@ -5,10 +5,11 @@ import Navigation from "@/components/Navigation";
 import { useTranslation } from 'react-i18next';
 import { LogOut, Users, MessageCircle } from 'lucide-react';
 import { useRouter } from 'next/navigation';
-import { IUser } from "@/entities/User";
-import { IMember } from "@/entities/Member";
-import { ISite } from "@/entities/Site";
 import { useLoginModalStore } from '@/store/LoginModalStore';
+import { useUserStore } from '@/store/UserStore';
+import { useMemberStore } from '@/store/MemberStore';
+import { useSiteStore } from '@/store/SiteStore';
+import { landingPage } from '@/app/settings';
 
 const LANGS = [
   { code: 'he', label: 'עברית', flag: '🇮🇱' },
@@ -16,7 +17,7 @@ const LANGS = [
   { code: 'tr', label: 'Türkçe', flag: '🇹🇷' },
 ];
 
-function getUserInitials(member?: IMember) {
+function getUserInitials(member?) {
   if (!member?.displayName) return 'U';
   return member.displayName
     .split(' ')
@@ -26,14 +27,7 @@ function getUserInitials(member?: IMember) {
     .slice(0, 2);
 }
 
-interface HeaderProps {
-  user?: IUser;
-  member?: IMember;
-  onLogout?: () => void;
-  siteInfo: ISite;
-}
-
-export default function Header({ user, member, onLogout, siteInfo }: HeaderProps) {
+export default function Header() {
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
   const [isLangMenuOpen, setIsLangMenuOpen] = useState(false);
   const [isUserMenuOpen, setIsUserMenuOpen] = useState(false);
@@ -41,6 +35,14 @@ export default function Header({ user, member, onLogout, siteInfo }: HeaderProps
   const userMenuRef = useRef(null);
   const router = useRouter();
   const openLogin = useLoginModalStore((s) => s.open);
+  const { user, logout } = useUserStore();
+  const member = useMemberStore((state) => state.member);
+  const siteInfo = useSiteStore((state) => state.siteInfo);
+
+  const handleLogout = async () => {
+    await logout();
+    router.push(landingPage);
+  };
 
   useEffect(() => {
     function handleClickOutside(event) {
@@ -69,12 +71,12 @@ export default function Header({ user, member, onLogout, siteInfo }: HeaderProps
     <header className="w-full flex items-center justify-between px-4 py-2 bg-white shadow-sm sticky top-0 z-50">
       {/* Left: Site title */}
       <div className="text-xl font-semibold text-sage-700">
-        {siteInfo.name}
+        {siteInfo?.name}
       </div>
       {/* Center: Navigation */}
       <div className="flex flex-row items-center">
-        {user && member && onLogout && member.role !== 'pending' && (
-          <Navigation user={user} onLogout={onLogout} setMobileMenuOpen={setMobileMenuOpen}/>
+        {user && member && member.role !== 'pending' && (
+          <Navigation user={user} onLogout={handleLogout} setMobileMenuOpen={setMobileMenuOpen}/>
         )}
       </div>
       {/* Right: Flags + Avatar */}
@@ -106,7 +108,7 @@ export default function Header({ user, member, onLogout, siteInfo }: HeaderProps
           )}
         </div>
         {/* Avatar + User Menu */}
-        {user && member && onLogout ? (
+        {user && member ? (
           <div className="hidden md:block relative" ref={userMenuRef}>
             <button
               onClick={() => setIsUserMenuOpen((v) => !v)}
@@ -157,7 +159,7 @@ export default function Header({ user, member, onLogout, siteInfo }: HeaderProps
                   <button
                     onClick={() => {
                       setIsUserMenuOpen(false);
-                      onLogout?.();
+                      handleLogout();
                     }}
                     className="flex items-center w-full px-4 py-2 text-sm text-sage-700 hover:bg-sage-50 transition-colors duration-200"
                   >
